Refactor: Split calculator.ts, optimize API sync, and cleanup logs
- Split `src/lib/calculator.ts` into modular files under `src/lib/calculators/`: `core.ts`, `stats.ts`, `charts.ts`.
- Preserved `src/lib/calculator.ts` as a facade to ensure backward compatibility.
- Optimized `src/routes/api/sync/orders/+server.ts`:
    - Added 50s timeout safety check to prevent serverless function timeouts.
    - Improved error propagation from Bitunix API.
    - Reduced max pages per batch for safety.
- Removed non-critical `console.log` statements from `bitunixWs.ts` and UI components to reduce noise.
- Verified with existing unit tests (`npm test` passed 121 tests).
